+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-api/src/main/java/com/forgerock/sapi/gateway/rcs/consent/store/api/v3_1_10/DomesticPaymentConsentApi.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.rcs.consent.store.api;
+package com.forgerock.sapi.gateway.rcs.consent.store.api.v3_1_10;
 
 import javax.validation.Valid;
